requirements.txt
pytest==7.2.0
selenium==4.6.0

python-dateutil~=2.8.2